Implement per-port type checking

+++ src/NodeEditor/Ports.cpp
@@ -5,6 +5,8 @@
 #include <imgui.h>
 #include <imnodes.h>
 
+#include <utility>
+
 static int gPortIDCounter = 1;
 
 Port::Port(const Node& parent, std::string displayName, int id) :
@@ -12,8 +14,8 @@ mParentNode(parent), mDisplayName(std::move(displayName)), mID{id == 0 ? gPortID
     gPortIDCounter = std::max(gPortIDCounter, mID + 1);
 }
 
-OutPort::OutPort(const Node& parent, const std::string& displayName, const std::any& value, int id) :
-Port(parent, displayName, id), mValue(value) {
+OutPort::OutPort(const Node& parent, const std::string& displayName, int id) :
+Port(parent, displayName, id) {
 }
 
 void OutPort::draw() const {
@@ -27,7 +29,7 @@ void OutPort::draw() const {
 }
 
 void OutPort::drawLinks() const {
-
+    // Link drawing handled by InPort
 }
 
 void OutPort::unlink() {
@@ -38,6 +40,8 @@ void OutPort::unlink() {
 }
 
 void OutPort::link(InPort& port) {
+    if (!port.isConnectionValid(*this))
+        return;
     if (mLink) {
         unlink();
     }
@@ -54,8 +58,9 @@ void OutPort::linkSoft(InPort& port) {
     mLink = &port;
 }
 
-InPort::InPort(const Node& parent, const std::string& displayName, int id) :
-Port(parent, displayName, id) {
+InPort::InPort(const Node& parent, const std::string& displayName,
+               std::vector<const std::type_info*> validConnections, int id) :
+Port(parent, displayName, id), mValidConnections(std::move(validConnections)) {
 }
 
 void InPort::draw() const {
@@ -79,6 +84,8 @@ void InPort::unlink() {
 }
 
 void InPort::link(OutPort& port) {
+    if (!isConnectionValid(port))
+        return;
     if (mLink) {
         unlink();
     }
@@ -94,4 +101,13 @@ void InPort::linkSoft(OutPort& port) {
     static int cLinkIDCounter = 1;
     mLink = &port;
     mLinkID = cLinkIDCounter++;
+}
+
+bool InPort::isConnectionValid(const IPort& port) {
+    if (mValidConnections.empty())
+        return true;
+    const Node& parent = port.getParent(); // Multiline to avoid typeid warning
+    const std::type_info* check = &typeid(parent);
+    return std::any_of(mValidConnections.begin(), mValidConnections.end(),
+                       [check](const std::type_info* a) { return *a == *check; });
 }
